Fixed some problems with the Pitzer model implementation.
The C^phi_MX parameter is given directly in tabulated data.
It does not have to be calculated or derived, but can be used directly
in the osmotic coefficient and activity coefficient equations. Previously,
the code was written under the assumption that the C_MX paramter
was given and the C^phi parameter had to be calculated.

The C_MX parameter is calculated from C^phi
(one half of C^phi for a univalent electrolyte), but does
not appear to be required in either of these two equations. As
such, the _pizter_C_phi() function was renamed to _pitzer_C_MX and
commented out for now.

For more, see:
May, P. M., Rowland, D., Hefter, G., & Königsberger, E. (2011).
A Generic and Updatable Pitzer Characterization of Aqueous Binary Electrolyte Solutions at 1 bar and 25 °C.
Journal of Chemical & Engineering Data, 56(12), 5066–5077. doi:10.1021/je2009329