Fix an error in concentration calculations when volume argument was specified. Dropped support for direct entry of solvent mass.
Previously, if the 'solvent' argument was left at its default value, the
water mass would be initialized to 0.998 kg irrespective of the volume
specified by the 'volume' argument. This led to incorrect concentration
calculations whenever a volume other than 1L was specified.

This commit rectifies this behavior by dropping support for directly
setting the solvent mass. Solvent mass is now calculated using the
total solution volume and the water density.

NOTE: this means that that 'volume' argument is really specifying the
solvent volume, not the total solution volume. This may be improved or
clarified later.