Re-enable setting of solvent mass to faciliate copy() method.
The previous commit disabled the ability to set the solvent mass
directly when initializing a solution, but this capability was needed
to make proper copies of solutions, since the solution mass changes
after solutes are added. This commit restores the ability to specify
a solvent and its mass, but adds an error message if both volume
and solvent mass are specified.