Iterative calculation of solute amounts with mol/L or similar units
This commit fixes the add_solute(), solution init, set_amount(), and
add_amount() methods so that if the user supplies an amount in
substance or mass per volume units (e.g. 'mol/L'), the required
quantity of moles is iteratively calculated so that
1) the total solution volume does not change
2) the displacement of solvent by solute is accounted for

Previously, the moles would be calculated based on the solution
volume. Then the volume would be updated to account for the new
solute. Thus, you could add '4 mol/L' of a solute and immediately
wind up with some lower concentration, because the solution
volume increased.

Also added a simple test script to the tests directory. This eventually
needs to be modified to allow automated testing.