Add pH support to Solution init
Due to a previous modification to the get_salt() method, it would return
a salt containing H+ or OH- in the event that no other solutes were found.
But, those solutes may not have actually been present in the solution,
leading to a key error.

This commit ensures that H+ and OH- are always present in any pyEQL
Solution.