Always use Pitzer activity model when parameters are available.
Previously, the Pitzer model was only invoked for ionic strength > 0.5.
While this had a speed advantage, it led to odd behavior when the starting
point for a simulation was a 0.5 mol/kg solution, because a small change
in composition would lead to a discontinuity in the activity.

With this commit, the Pitzer model will be used to calculate activity
coefficients whenever the parameters are available, regardless of ionic strength.