Fix bug in get_activity_coefficient()
Previously, the code would search for pitzer parameters and flip
the found variable if item was equal to the pitzer parameters. However,
it would continue searching. Then, item would be passed to the pitzer
model as an argument. In some cases, item would not necessarily be
the expected pitzer paramters, leading to an index error.

This commit assures that pyEQL stops searching the database as soon
as pitzer parameters are found, so that the activity correction function
gets the expected arguments.