rewrite add_amount(), set_amount(), and _update_volume() for performance
This commit rewrites the above functions in a way that avoids the need
for iterative recalculation of concentration when molar units are
specified. It preserves the logic from before -- if mol/kg (or equivalent)
units are used, the solution volume is recalculated and that's it.

On
the other hand, if mol/L (or equivalent) units are used, the volume is held
constant and the mass of SOLVENT is modified to make room for the additional
solute. The old method of handling this case arbitrarily reset the moles of
each (other) solute in order to preserve their original concentrations, but
didn't consider the impact of this change on the total solution volume.

So the new way is faster and more accurate.