Defer recalculation of solution volume until requested by get_volume()
Previously, the solution volume would be recalculated every time a solute
was added/removed. Because this update can involve calculations using the
Pitzer model, it can be somewhat expensive in terms of performance.

This update replaces calles to _update_volume with a volume_update_required
flag. If the flag is set to True, then the volume update will be triggered
the next time the get_volume() method is called.

Implementing this feature required changes to get_amount() to prevent an
infinite loop (_update_volume calls get_amount, which originally called
get_volume() even when mol/kg units were used).

This commit also updates the code in add_solute() to match recent changes
to add_amount, avoiding the need for an iterative calculation when
mol/L units are used and improving the accuracy.