Implement Erying-type equation for viscosity calculation
This equation is supposedly applicable over a wider range of salt concentrations
than the Jones-Dole method used before. This one calculates the
kinematic viscosity directly, so the logic of the dynamic and relative
viscosity functions had to be altered so that they are all computed
based on the kinematic viscosity.

The new method looks for parameters named 'erying_viscosity_coefficients'
in the database.