Add new logging feature to functions.py and water_properties.py Make donnan_eql() tolerant of a wider range of partial molar volume cases.
Previously, donnan_eql() required that partial molar volumes were provided
for each ion in the salt; even if a value for the overall salt was in
the database. This commit allows the function to accept a partial molar
volume for the entire salt in place of individual ionic contributions, and
logs an error if neither is available.