Invoke the pitzer model only for the right cations / anions
Previously, in a multi electrolyte solution, if get_activity() was
called for an ion that was not in the primary salt (say, H+ in a
solution of NaCl), pyEQL would still use the Pitzer model for NaCl
to calculate its activity coefficient.

This commit fixes the behavior so that the Pitzer model is only invoked
for ions that correspond to the model parameters.