Solution: change averaging method for osmotic coefficient
Averaging the osmotic coefficients for individual salts based on their
effective molality (as presented in the Mistry et al. paper) was causing
pyEQL to fail tests for single electrolyte solutions, and made little
conceptual sense. This commit revises the averaging to use the actual
molality of each salt, and results in the "Effective Pitzer" model
agreeing with the regular Pitzer model for single electrolyte solutions.