Update test suite to test relative error instead of decimal places
This commit replaces assertAlmostEqual with
assertWithinExperimentalError for all tests in which pyEQL output is
compared to experimental data. The tolerable relative error is specified
in the setUp method of each test class.

After making this change, the number of tests expected to fail was
reduced from 4 to 2, and all remaining tests predict experimental data
within 1-5%