Implement function to calculate mean activity
Although EOS subclasses define get_activity_coefficicient to return the
activity coefficients of individual solutes, one can only measure the
mean activity coefficient experimentally. As such, to benchmark against
experimental data, we must use the definition of the mean activity
coefficient in terms of the activity coefficients of the individual ions.
Alternatively, this function could be implemented as a property
(Solution.mean_activity) or a method (Solution.get_mean_activity).

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>