Fix property retrieval, RMSE calc; rename vars
The new RMSE is more robust in the case that property values are Python
primitives.

Property getters are now called to retrieve properties. This fixes a bug
where the getter was returned instead of the value.

Variables are renamed (from property) to avoid shadowing a Python built-in.

Enumerating solution data (in get_dataset) yields the right number of items
to unpack.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>