Defer decision on key structure
Encapsulating the logic for creating dictionary keys in a function and creating
a type alias for its type should minimize the amount of refactor required if the
format is changed and appease mypy. This also makes comparison of Solutions
for the purpose of benchmarking considerably more simple.

load_dataset now returns a dictionary with keys indicating the Solution for which
the data applies as generated by _create_solution_key.

Internal data sources are now opened more reliably using resources.files.

Generating the list of SolutionKeys corresponding to the `solutions` parameter in advance
simplifies the logic required to check for the Solution in `solutions`.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>