Use same data structures for engine/reference data
Using the same data structures for engine and reference data is desired because
it will separate the logic for retrieving properties and calculating stats and
simplify the retrieval of corresponding items in calculate_stats.

This implementation ensures that only one Solution object is created for each
Solution across all benchmark datasets and that each BenchmarkEntry object is
associated with a unique Solution. In this way, a single dataset can be generated
for the engine, which can then be benchmarked against each reference dataset.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>