Implement reference/calculated for calculate_stats
Accepting two identical data structures generalizes calculate_stats for
all combinations of sources of data (e.g., reference-reference, engine-engine,
engine-reference). An additional benefit is that now, from the signature, it
is more clear to users how to organize the two sets of data to be compared.
Previously, engine data was implicitly generated using the Solution objects in
each BenchmarkEntry. As mentioned previously, this now removes the logic for
retrieving solute and solution properties.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>