Remove use of FormulaDict; fix bugs
FormulaDict was initially used in order to automatically standardize how solutes
appear in solution compositions and solute property lists. However, in addition
to standardizing ion formulas, FormulaDict also sorts its internal copy of the
dictionary by its values. This led to an error in the previous implementation
because the values are dictionaries, and dictionaries don't implement the required
methods for comparison with '<' and '>'.

Instead, we manually call standardize_formula in the public functions where
malformatted formulas are likely. This is when loading datasets and when
creating a BenchmarkEntry from a Solution.

Bugs: initialize property data as dictionaries in create_entry; fix indices and
data type in _create_engine_dataset; initialize data pair lists in calculate_stats.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>