Update docstrings; Fix metric args
Updated module docstring variable for consistency.

Require both _create_engine_dataset arguments.

For now, only require that the metric argument to calculate_stats can handle
Quantity arguments, so that we can ensure that units are handled correctly.
Note that the property values in BenchmarkEntry.solute_data and
BenchmarkEntry.solution_data are expected to be Quantity objects.

Also, the additional conditional statements in calculate_stats serve to handle
the edge cases of when the return values of _get_solute_property and
_get_solution_property are None and when there does not exist data in `calculated`
corresponding to `reference`. Because this may be of intereste to the user, the
event is logged at the INFO level.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>