Use newest ver. of Python 3.10-compatible IPython
IPython dropped support for Python 3.10 after version 8.37.0 which was causing the
error in CI.

Notet that since the docs are built using Python 3.10, this limits the newest version of
IPython that can be used to 8.37.0. It's worth noting, however, that I am able
to compile the docs on ARM MacOS in a Python 3.11.9 environment. Doing so reduces
the number of warnings from 24 to 3. To implement this change, one would need to:

- change the value of the version of Python that is used in the docs job
(`testing.yaml`) in the setup-python@v5 GH action to "3.11"
- pin the IPython version in `docs/requirements.txt` using ipython>=9.3.0
- set build.tools.python to "3.11" in `.readthedocs.yml`

Signed-off-by: Ugochukwu Nwosu <ugn@sfu.ca>