Make pyEQL compatible with Solution.get_salt_dict
Note that there is also a change in the logic of NativeEOS.get_activity_coefficient.
Previously, when the for loop iterated over solution.get_salt_dict().values(),
the check `v == 'HOH'` was used. However, this condition would always evaluate to
false as v would have been a dictionary. This patch changes the conditional expression
to correctly compare the formula of the salt to 'HOH'. A regression test should also
be included.

Signed-off-by: Ugochukwu Nwosu <ugn@sfu.ca>