Remove components unit tests; loosen type check
In a previous commit (5fa5a319), the type checks on Solution.components
values and the values corresponding to the 'mol' keys in the return value
of Solution.get_salt_dict() were made stricter to fail with subclasses of float.
This was in anticipation of correcting a minor bug originating from the use of
numpy.float64's in certain functions (e.g., Solution.water_substance.rho,
activity_correction.get_apparent_volume_pitzer). For example,

>>> from pyEQL import Solution
>>> sol = Solution()
>>> sol.components
{'H2O(aq)': np.float64(55.34457595832858), 'OH[-1]': 1.0000000000000001e-07, 'H[+1]': 1e-07}
>>> sol.get_salt_dict()
{'HOH': {'salt': <pyEQL.salt_ion_match.Salt object at 0x111f32590>, 'mol': np.float64(55.34457595832858)}}

That is, the number of moles of the solvent is stored as a numpy.float64, but the number of moles of any
other component is stored as an ordinary Python float. This is somewhat harmless since
numpy.float64's are subclasses of Python floats, but it might be unexpected. In any case,
correcting this behaviour is beyond the scope of the current changes and requires
changes in many modules, so the unit tests checking for this behaviour have been removed.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>