Consolidate/expand get_salt_dict tests
A number of unit tests for Solution.get_salt_dict are introduced to improve
method coverage to 100%. Tests are parametrized to cover combinations of
uni-, di-, and trivalent ions as well as empty solutions.

test_should_not_include_water is tentatively included pending the proposed
change to Solution.get_salt_dict (i.e., ignore water).

The following tests are failing due to bugs in the code:
- test_should_order_salts_by_amount
- test_should_not_return_salts_with_concentration_below_cutoff

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>