Annotate salt_ratio fixture
salt_ratio must be set so that .get_salt_dict is able to correctly pair
anions and cations. Without such a condition, it is not possible, even in
principle, to identify the salts in an arbitrary mixture of ions. For example,
consider a solution for which the concentrations of Na[+1], K[+1], Cl[-1], and
SO4[-2] are 1 M, 1.5 M, 1 M, and 0.75 M, respectively. One can associate with this
1 mole of NaCl and 0.75 moles of K2SO4 or equivalently 1 mole of
KCl, 0.25 moles K2SO4, and 0.5 moles Na2SO4.

To avoid this one must set salt_ratio such that for any two successive
cations/anions in `salts`, salt_ratio * nu_ratio < 1.0, where nu_ratio is the
ratio of the stoichiometric coefficient of the cation/anion that appears later in
`salts` to that of the *preceding* Salt. For a solution of NaCl and K2SO4
(with [NaCl] > [K2SO4]) to be unambiguously identified, salt_ratio must be less than
0.5 (i.e., because nu_cation > 2.0, we must have [NaCl] / [K2SO4] > 0.5).

Signed-off-by: Ugochukwu Nwosu <ugn@sfu.ca>