Reduce parametrization for multisalt solutions
Note that although the new parametrization covers fewer test cases, it
still covers mono-, di-, and trivalent ions, and some character of the
removed tests remains. For example, _SOLUTES_LITE won't contain a salt-ion
pair corresponding to FeCl3, but it will contain a salt-ion pair corresponding
to Fe(OH)3-another salt featuring a trivalent cation with a monovalent
anion. Further, the combination of Fe(OH)3 with HCl will result in a solution
featuring both Fe[+3] and Cl[-1] ions (although not in stoichiometric ratios).

Signed-off-by: Ugochukwu Nwosu <ugn@sfu.ca>