Fix salt concentration bug with polyvalent ions
Previously, when matching ions to form salts in Solution.get_salt_dict,
if the cation equivalents did not exceed the anion equivalents and the
cation was polyvalent (i.e., nu_cation > 1), then the concentration of
the salt was incorrectly calculated due to missing parentheses.

This bug was identified by the test_should_calculate_correct_concentration_for_salts
unit test.

Signed-off-by: Ugochukwu Nwosu <ugn@sfu.ca>