Implement cutoff in units of moles
Calling Solution.volume in Solution.get_salt_dict results in a RecursionError
because if isinstance(Solution.engine, NativeEOS), Solution.volume calls
Solution.get_salt_dict.

Thus, the only way that the cutoff parameter can be implemented is if it
is interpreted as the nominal amount of salt that must be present for
it to be included in the output of get_salt_dict. Although this is a less
than ideal interface, robust use is straightforward:

>>> sol = Solution(solutes={...})
>>> amount = ureg.Quantity("0.5 mol/L") * sol.volume
>>> list(sol.get_salt_dict(cutoff=amount.m))
'NaCl'

Signed-off-by: Ugochukwu Nwosu <ugn@sfu.ca>