Fix bug related to adding solutions + unit tests
Sum of solutions does not inherit their engines/solvent/database
-----------------------------------------------------------------

When adding two solutions, the engine of the parent solutions are not passed
when creating the new solution. This was identified by the third assert statement in
test_mixing_functions (i.e., `assert np.isclose(gibbs_mix(s1_p, s1_p).magnitude, 0)`).
This test failed because the engine of the summand solutions was "phreeqc", but
since no engine is passed when creating the resulting sum, its engine defaults to
native. As a result, Solution.get_activity returns different values for the activities
of H+/OH- in the blend and the Gibbs energy of mixing is non-zero.

The fix includes parametrized unit tests checking that the volume, engine, atomic
amounts, solvent, and database are conserved.

NOTE: The refactoring in test_conductivity was required to prevent reformatting by
Black and Ruff simultaneous while running pre-commit.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>