Concretize new NativeEOS method behaviour w/ tests
Due to the recent changes to Solution.get_salt_dict,
there are subtle changes to how the `NativeEOS` methods execute.
This commit introduces unit tests to clarify and preserve
the new behaviour.

Consider the following snippet:

```
from pyEQL import Solution
sol1 = Solution()
sol1.get_activity_coefficient('H+').m
sol2 = Solution(solutes={"Na[+1]": "1 mol/L"})
sol2.get_osmotic_coefficient().m
sol2._get_solute_volume().m
```

and its output before and after the recent changes:

```
np.float64(0.9996283313178875)
np.float64(0.8502757138102266)
np.float64(-0.0015974709364873756)
1
1.0
-0.0050000004
```

Previously, H+ and OH- activity coefficients were calculated
according to the ionic strength of the solution. (This is a
result of the `if v == 'HOH'` condition never evaluating to True.)
Now, the activity coefficients of H+ and OH- will be unity unless
they are a component in a salt whose concentration exceeds
default cutoff used by `Solution.get_salt_dict`. In fact, **the
activity coefficient of any solute will be unity unless it is a
component in a salt whose concentration exceeds the default cutoff.**

Contributions to the average osmotic coefficient returned by
`NativeEOS.get_osmotic_coefficient` are determined by the return value
of `Solution.get_salt_dict`. Thus, previously, all components that
were part of a salt whose formula was not 'HOH' contributed to the
average osmotic coefficient. (This was because the cutoff parameter
was not respected by Solution.get_salt_dict.) Now, only salts with
concentrations above the default cutoff will contribute to the
osmostic coefficient.  In the snippet above, the osmotic coefficient
of 0.8502757138102266 in the previous implementation is that of 1e-7 M
NaOH.

In calculating the solute volume, an attempt is made to first
calculate a solute volume based on the apparent volume of the major
salt in the Solution using its Pitzer parameters. Then, solute volumes
are calculated for the remaining non-solvent components using internal
molar volume parameters. Because the cutoff parameter was previously
not respected, the call to `Solution.get_salt()` returned an 'NaOH'
Salt object, and the contributions of Na+ and OH- to the molar volume
were determined by the apparent volume of NaOH. Now, because the
concentration of NaOH (1e-7 M) is below the default cutoff used by
Solution.get_salt_dict, None is returned by `Solution.get_salt()`, and
the contributions of Na+ and OH- to the molar volume are determined
from the internal database.

Note: This comment contradicts statements made in a previous commit
(73b1261b6c20adec2b5b90fa962aa64399d54c25), as the case considered here was
not considered previously.

Unit tests have been added to explicitly outline/preserve the current behaviour.

Signed-off-by: Ugochukwu Nwosu <ugognw@gmail.com>